Turn a gint16 into gint
Because, I mean, really.

+++ gtk/gtkpathbar.c
@@ -92,7 +92,7 @@ struct _GtkPathBarPrivate
   GtkWidget *down_slider_button;
   guint settings_signal_id;
   gint icon_size;
-  gint16 slider_width;
+  gint slider_width;
   guint timer;
   guint slider_visible : 1;
   guint need_timer     : 1;
